FASTAPI/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
pydantic
email-validator
sqlalchemy[asyncio]
asyncpg
aiofiles
redis
python-dotenv
passlib[bcrypt]
aiohttp
python-socketio
pillow
fastapi-mail
